==> app_SOWmAcHiNe_2019/1_main.ipynb <==
| MAIN BASELINE TEMPLATE | Create baseline Statement of Work (SOW) Excel template |
| ----------- | ----------- |

import xlwings as xw
from openpyxl import Workbook, load_workbook
import xlsxwriter 

import pandas as pd
import numpy as np
import datetime as dt 
import time
import sys

"""
Latest xlwings release: v0.27.15
https://www.xlwings.org/

"""
xw.__version__ 

from appscript import k

# Connect to unsaved book (looks in all Excel instances)
wb = xw.Book()

# Create a sheet object (and reference first sheet)
sheet = wb.sheets[0] # Uses standard python indexing

sow_header_list = ['stage',
                   'task',
                   'task_details',
                   'role',
                   'hours',
                   'rate',
                   'cost']

# print(sow_header_list)
# print(len(sow_header_list))

"""
This is my 'Hello World'

"""

# Datetime (start)
sheet.range('H2').value = dt.datetime.now()

# Add space in column to be able to see the time element and not ######## I am using as the "Hello World"
sheet.range('H:H').column_width = 15

# Index notation (1-based like excel)
sheet.range((2,8)).value # Verify date in both excel and python terminal or jupyter output

# Create SOW header from list
sheet.range((1,1)).value = sow_header_list # Horizontal (row)
sheet.range('A1:G1').font.bold = True

# Set global named range
sheet.range('A1:G500').name = 'sow_nr'

sow_brd_loc = sheet['sow_nr'][0:179,:]

# Set border around excel SOW
sow_brd_loc[-1,:].api.get_border(which_border=k.border_bottom).weight.set(4)
sow_brd_loc[:,-1].api.get_border(which_border=k.border_right).weight.set(4)
sow_brd_loc[0,:].api.get_border(which_border=k.border_top).weight.set(4)
sow_brd_loc[:,0].api.get_border(which_border=k.border_left).weight.set(4)

# Set phases
sheet['sow_nr'][1,0].value = 'pre-kickoff'
sheet['sow_nr'][19,0].value = 'inititiate'
sheet['sow_nr'][39,0].value = 'define'
sheet['sow_nr'][79,0].value = 'design'
sheet['sow_nr'][99,0].value = 'build'
sheet['sow_nr'][139,0].value = 'transition'
sheet['sow_nr'][159,0].value = 'close'

# Set phases
sheet['sow_nr'][1,0].font.bold = True
sheet['sow_nr'][19,0].font.bold = True
sheet['sow_nr'][39,0].font.bold = True
sheet['sow_nr'][79,0].font.bold = True
sheet['sow_nr'][99,0].font.bold = True
sheet['sow_nr'][139,0].font.bold = True
sheet['sow_nr'][159,0].font.bold = True

# Set phase named ranges
sheet['A2:G19'].name = 'env_nr'
sheet['A20:G39'].name = 'int_nr'
sheet['A40:G79'].name = 'def_nr'
sheet['A80:G99'].name = 'des_nr'
sheet['A100:G139'].name = 'bld_nr'
sheet['A140:G159'].name = 'trn_nr'
sheet['A160:G179'].name = 'cls_nr'

# Set phase divider color
sheet['env_nr'][0,:].color = '#E21A23'
sheet['int_nr'][0,:].color = '#1C0221'
sheet['def_nr'][0,:].color = '#440850'
sheet['des_nr'][0,:].color = '#880074'
sheet['bld_nr'][0,:].color = '#6C0BA9'
sheet['trn_nr'][0,:].color = '#9D85F3'
sheet['cls_nr'][0,:].color = '#86EC77'

# Set phase font color
sheet['env_nr'][0,0].font.color = '#ffffff'
sheet['int_nr'][0,0].font.color = '#ffffff'
sheet['def_nr'][0,0].font.color = '#ffffff'
sheet['des_nr'][0,0].font.color = '#ffffff'
sheet['bld_nr'][0,0].font.color = '#ffffff'
sheet['trn_nr'][0,0].font.color = '#000000'
sheet['cls_nr'][0,0].font.color = '#000000'

# Set wrap text
# sheet.range('A:A').wrap_text = True
sheet.range('B:B').wrap_text = True
sheet.range('C:C').wrap_text = True 
sheet.range('D:D').wrap_text = True 
sheet.range('F:F').wrap_text = True

# Set column width
sheet.range('A:A').column_width = 11
sheet.range('B:B').column_width = 22
sheet.range('C:C').column_width = 22*3
sheet.range('D:D').column_width = 15
sheet.range('E:E').column_width = 15
sheet.range('F:F').column_width = 15
sheet.range('G:G').column_width = 15
# sheet.range('H:H').column_width = 15 # This is the time element outside of table

# Set phase bottom divider color
sheet['env_nr'][-2:,:].color = '#E21A23'
sheet['int_nr'][-2:,:].color = '#1C0221'
sheet['def_nr'][-2:,:].color = '#440850'
sheet['des_nr'][-2:,:].color = '#880074'
sheet['bld_nr'][-2:,:].color = '#6C0BA9'
sheet['trn_nr'][-2:,:].color = '#9D85F3'
sheet['cls_nr'][-2:,:].color = '#86EC77'

# Store hours and cost total list as variables
totals_stage_list = ['Hours', 0, 'Cost', 0]
totals_cltv_list = ['Hours (Agg.)', 0, 'Cost (Agg.)', 0]

"""
One Formula and Additional formatting
"""

"""
Envision
"""
# Cost (hrs*rate)
sheet['env_nr'][1:-2, 6].value = '$0.00' # Cost column format
sheet['env_nr'][1:-2, 5].value = '$0.00' # Rate column format
sheet['env_nr'][1:-2, 6].formula = '=(E3*F3)'
# Totals
sheet['env_nr'][-2, 3].value = totals_stage_list
sheet['env_nr'][-1, 3].value = totals_cltv_list
# Font color
sheet['env_nr'][-2:,3:].font.bold = True
sheet['env_nr'][-2:,3:].font.color = '#ffffff'

""" 
Initiate
"""

# Cost (hrs*rate)
sheet['int_nr'][1:-2, 6].value = '$0.00'
sheet['int_nr'][1:-2, 5].value = '$0.00'
sheet['int_nr'][1:-2, 6].formula = '=(E21*F21)'
# Totals
sheet['int_nr'][-2, 3].value = totals_stage_list
sheet['int_nr'][-1, 3].value = totals_cltv_list
# Font color
sheet['int_nr'][-2:,3:].font.bold = True
sheet['int_nr'][-2:,3:].font.color = '#ffffff'

""" 
Define
"""

# Cost (hrs*rate)
sheet['def_nr'][1:-2, 6].value = '$0.00'
sheet['def_nr'][1:-2, 5].value = '$0.00'
sheet['def_nr'][1:-2, 6].formula = '=(E41*F41)'
# Totals
sheet['def_nr'][-2, 3].value = totals_stage_list
sheet['def_nr'][-1, 3].value = totals_cltv_list
# Font color
sheet['def_nr'][-2:,3:].font.bold = True
sheet['def_nr'][-2:,3:].font.color = '#ffffff'

""" 
Design
"""
# Cost (hrs*rate)
sheet['des_nr'][1:-2, 6].value = '$0.00'
sheet['des_nr'][1:-2, 5].value = '$0.00'
sheet['des_nr'][1:-2, 6].formula = '=(E81*F81)'
# Totals
sheet['des_nr'][-2, 3].value = totals_stage_list
sheet['des_nr'][-1, 3].value = totals_cltv_list
# Font color
sheet['des_nr'][-2:,3:].font.bold = True
sheet['des_nr'][-2:,3:].font.color = '#ffffff'

""" 
Build
"""
# Cost (hrs*rate)
sheet['bld_nr'][1:-2, 6].value = '$0.00'
sheet['bld_nr'][1:-2, 5].value = '$0.00'
sheet['bld_nr'][1:-2, 6].formula = '=(E101*F101)'
# Totals
sheet['bld_nr'][-2, 3].value = totals_stage_list
sheet['bld_nr'][-1, 3].value = totals_cltv_list
# Font color
sheet['bld_nr'][-2:,3:].font.bold = True
sheet['bld_nr'][-2:,3:].font.color = '#ffffff'

""" 
Transition
"""
# Cost (hrs*rate)
sheet['trn_nr'][1:-2, 6].value = '$0.00'
sheet['trn_nr'][1:-2, 5].value = '$0.00'
sheet['trn_nr'][1:-2, 6].formula = '=(E141*F141)'
# Totals
sheet['trn_nr'][-2, 3].value = totals_stage_list
sheet['trn_nr'][-1, 3].value = totals_cltv_list
# Font color
sheet['trn_nr'][-2:,3:].font.bold = True
sheet['trn_nr'][-2:,3:].font.color = '#000000'

""" 
Close
"""
# Cost (hrs*rate)
sheet['cls_nr'][1:18, 6].value = '$0.00'
sheet['cls_nr'][1:18, 5].value = '$0.00'
sheet['cls_nr'][1:18, 6].formula = '=(E161*F161)'
# Totals
sheet['cls_nr'][18, 3].value = totals_stage_list
sheet['cls_nr'][19, 3].value = totals_cltv_list
#Font color
sheet['cls_nr'][18:,3:].font.bold = True
sheet['cls_nr'][18:,3:].font.color = '#000000'


"""
For each phase (aka phase named range)
Column G sum Cost
Column G sum Cost Agg.

Column E sum Hours
Column E sum Hours Agg.

Add left column A thick boarder by project stage

"""

"""
Envision
"""
# Hours
sheet['env_nr'][-2,4].formula = '=SUM(E3:E17)'
sheet['env_nr'][-1,4].value = '=E18'
# Cost
sheet['env_nr'][-2,6].formula = '=SUM(G3:G17)'
sheet['env_nr'][-1,6].formula = '=G18' # Value for G19
# Boarder
sheet['env_nr'][:,0].color = '#E21A23'

"""
Initiate
"""
# Hours
sheet['int_nr'][-2,4].formula = '=SUM(E21:E37)'
sheet['int_nr'][-1,4].formula = '=SUM(E38+E19)'
# Cost
sheet['int_nr'][-2,6].formula = '=SUM(G21:G37)'
sheet['int_nr'][-1,6].formula = '=SUM(G38+G19)'
# Boarder
sheet['int_nr'][:,0].color = '#1C0221'

"""
Define
"""
# Hours
sheet['def_nr'][-2,4].formula = '=ROUNDUP(SUM(E41:E77),0)'
sheet['def_nr'][-1,4].formula = '=ROUNDUP(SUM(E78+E39),0)'
# Cost
sheet['def_nr'][-2,6].formula = '=ROUNDUP(SUM(G41:G77),0)'
sheet['def_nr'][-1,6].formula = '=ROUNDUP(SUM(G78+G39),0)' # Careful with (parantheses)
# Boarder
sheet['def_nr'][:,0].color = '#440850'

"""
Design
"""
# Hours
sheet['des_nr'][-2,4].formula = '=SUM(E81:E97)'
sheet['des_nr'][-1,4].formula = '=SUM(E98+E79)'
# Cost
sheet['des_nr'][-2,6].formula = '=SUM(G81:G97)'
sheet['des_nr'][-1,6].formula = '=SUM(G98+G79)'
# Boarder
sheet['des_nr'][:,0].color = '#880074'

"""
Build
"""
# Hours
sheet['bld_nr'][-2,4].formula = '=SUM(E101:E137)'
sheet['bld_nr'][-1,4].formula = '=SUM(E138+E99)'
# Cost
sheet['bld_nr'][-2,6].formula = '=SUM(G101:G137)'
sheet['bld_nr'][-1,6].formula = '=SUM(G138+G99)'
# Boarder
sheet['bld_nr'][:,0].color = '#6C0BA9'

"""
Transition
"""
# Hours
sheet['trn_nr'][-2,4].formula = '=SUM(E141:E157)'
sheet['trn_nr'][-1,4].formula = '=SUM(E158+E139)'
# Cost
sheet['trn_nr'][-2,6].formula = '=SUM(G141:G157)'
sheet['trn_nr'][-1,6].formula = '=SUM(G158+G139)'
# Boarder
sheet['trn_nr'][:,0].color = '#9D85F3'

"""
Close
"""
# Hours
sheet['cls_nr'][-2,4].formula = '=SUM(E161:E177)'
sheet['cls_nr'][-1,4].formula = '=SUM(E178+E159)'
# Cost
sheet['cls_nr'][-2,6].formula = '=SUM(G161:G177)'
sheet['cls_nr'][-1,6].formula = '=SUM(G178+G159)'
# Boarder
sheet['cls_nr'][:,0].color = '#86EC77'

# Assign values to cells by passing lists to the excel doc
sheet['sow_nr'][179,3:].value = ['TOTAL HOURS', 0, 'TOTAL COST', '$0.00']
sheet['sow_nr'][180,3:].value = ['', 'OTD', '0%', '$0.00']
sheet['sow_nr'][181,3:].value = ['', '', 'GRAND TOTAL', '$0.00']

# Set font bold and color
sheet['sow_nr'][179:182,:].font.bold = True
sheet['sow_nr'][179:182,:].font.color = '#000000'

tot_brd_loc = sheet['sow_nr'][179:182,3:]

# Set border around excel SOW
tot_brd_loc[-1,:].api.get_border(which_border=k.border_bottom).weight.set(4)
tot_brd_loc[:,-1].api.get_border(which_border=k.border_right).weight.set(4)
tot_brd_loc[0,:].api.get_border(which_border=k.border_top).weight.set(4) # Not necessary really
tot_brd_loc[:,0].api.get_border(which_border=k.border_left).weight.set(4)


# Set formulas
tot_brd_loc[0,1].formula = '=(E179)'
tot_brd_loc[0,3].formula = '=(G179)'
tot_brd_loc[1,3].formula = '=(F181*G180)'
tot_brd_loc[2,3].formula = '=(G180-G181)'

"""
Build a math checker
-Do something later like IF 'GRAND TOTAL' == 'GRAND TOTAL' THEN font.color=green ELSE font.color=red
"""

# Set a math checker named range from the sow named range
math_chk_loc = sheet['sow_nr'][183:185,3:]

# Assign values to cells by passing lists to the excel doc
math_chk_loc[0,0].value = 'MATH CHECKER'
math_chk_loc[1,:].value = ['TOTAL HOURS', 0, 'TOTAL COST', '$0.00']

# Set color on math checker field
math_chk_loc[0,0].font.color = '#E21A23'

# Set font bold and color
math_chk_loc[0:2,:].font.bold = True
math_chk_loc[1:,:].font.color = '#000000'

# Set formulas
math_chk_loc[1,1].formula = '=SUM(E18+E38+E78+E98+E138+E158+E178)' # Total hours
math_chk_loc[1,3].formula = '=SUM(G18+G38+G78+G98+G138+G158+G178)-G181' # Total cost

# Set border and line style
math_chk_loc[1:2,:].api.get_border(which_border=k.border_top).weight.set(4)
math_chk_loc[1:2,:].api.get_border(which_border=k.border_left).weight.set(4)
math_chk_loc[1:2,:].api.get_border(which_border=k.border_bottom).weight.set(4)
math_chk_loc[1:2,:].api.get_border(which_border=k.border_right).weight.set(4)

# set roles and rates named range
role_rate_loc = sheet['sow_nr'][187:200,:]

# Assign header value in excel
rr_header_list = ['ROLE', 
                  'ROLE NAME', 
                  'ROLE DESCRIPTION', 
                  'STD RATE', 
                  'DISCOUNT', 
                  'DISCOUNT %', 
                  'RATE']

role_rate_loc.value = rr_header_list

# Read in ROLES_RATES df 
rr_df = pd.read_excel('sow_input_data.xlsx', sheet_name = 'ROLES_RATES')

rr_df = rr_df.iloc[0:7,:] # Filters out TAM role (for managed services)

rr_df

# Insert rows rates dataframe 
role_rate_loc.options(index=False, header=True).value = rr_df # IMPORTANT header=True

# DEV - insert the PS (pre-sales placeholder) role
ps_ph_list = ['PS', 'pre-sales', 'placeholder', '4', '0', '0', '0']

# insert ps_ph list for DEV
role_rate_loc[8,:].value = ps_ph_list
role_rate_loc[8,:].font.color = '#E21A23'

# Set formula to calculate final rate from discount (discount applied manually by user)
role_rate_loc[1:9,6].formula = '=(D189-E189)'

# DEV
# role_rate_loc[-1,6].formula = '=(D200-E200)'

"""
Set discount formula
-Important for usability by non-technical user
""" 

# Set formula to calculate discount perecentage
role_rate_loc[1:9,5].formula = '=(E189/D189)'

""" 
This looks sick from a UX perspective if the user is watching      
"""
# Set 0% and $0.00 formatting
role_rate_loc[1:,3].number_format = '$0.00' # STD RATE
role_rate_loc[1:,4].number_format = '$0.00' # DISCOUNT
role_rate_loc[1:,5].number_format = '0.0%' # DISCOUNT %
role_rate_loc[1:,6].number_format = '$0.00' # NEGOTIATED RATE

""" 
Format roles and rates table in excel doc
"""
# Set bold header
role_rate_loc[0,:].font.bold = True

# Set border and line style
role_rate_loc[0:8,:].api.get_border(which_border=k.border_top).weight.set(4)
role_rate_loc[0:8,:].api.get_border(which_border=k.border_left).weight.set(4)
role_rate_loc[0:8,:].api.get_border(which_border=k.border_bottom).weight.set(4)
role_rate_loc[0:8,:].api.get_border(which_border=k.border_right).weight.set(4)

"""
PLACEHOLDERS
"""
# for ROLE table append line to use in DEV
ps_role = ['PS', 'pre-sales', 'placeholder', '$0.00', '$0.00', '0.0%', '$0.00']

# string to insert in each row
# ps_placeholder = ['pre-sales', 'placeholder', 'PS', '4']
ps_placeholder = ['XXX', 'XXX', 'PS', '4']

"""   
Insert PLACEHOLDERS
"""
# insert placeholder lists in excel
sheet['env_nr'][1:,1:5].value = ps_placeholder
sheet['int_nr'][1:,1:5].value = ps_placeholder
sheet['def_nr'][1:,1:5].value = ps_placeholder
sheet['des_nr'][1:,1:5].value = ps_placeholder
sheet['bld_nr'][1:,1:5].value = ps_placeholder
sheet['trn_nr'][1:,1:5].value = ps_placeholder
sheet['cls_nr'][1:,1:5].value = ps_placeholder

# COPY / PASTE placeholder string from excel (at each stage)

# ENVISTION (PRE-KICKOFF)
sheet['env_nr'][1,1:].copy()
sheet['env_nr'][2:-2,1:].paste()

# INITIATE
sheet['int_nr'][1,1:].copy()
sheet['int_nr'][2:-2,1:].paste()

# DEFINE
sheet['def_nr'][1,1:].copy()
sheet['def_nr'][2:-2,1:].paste()

#DESIGN
sheet['des_nr'][1,1:].copy()
sheet['des_nr'][2:-2,1:].paste()

#BUILD
sheet['bld_nr'][1,1:].copy()
sheet['bld_nr'][2:-2,1:].paste()

#TRANSITION
sheet['trn_nr'][1,1:].copy()
sheet['trn_nr'][2:-2,1:].paste()

#CLOSE
sheet['cls_nr'][1,1:].copy()
sheet['cls_nr'][2:-2,1:].paste()

""" 
VLOOKUP() to populate rates for roles
-review everything about cell addresses before proceeding
"""
# ENVISION (PRE-KICKOFF)
sheet['env_nr'][1:-2, 5].formula = '=VLOOKUP(D3,$A$188:$G$196,7,FALSE)'

# INITIATE
sheet['int_nr'][1:-2, 5].formula = '=VLOOKUP(D21,$A$188:$G$196,7,FALSE)'

# DEFINE
sheet['def_nr'][1:-2, 5].formula = '=VLOOKUP(D41,$A$188:$G$196,7,FALSE)'

# DESIGN 
sheet['des_nr'][1:-2, 5].formula = '=VLOOKUP(D81,$A$188:$G$196,7,FALSE)'

# BUILD
sheet['bld_nr'][1:-2, 5].formula = '=VLOOKUP(D101,$A$188:$G$196,7,FALSE)'

# TRANSITION
sheet['trn_nr'][1:-2, 5].formula = '=VLOOKUP(D141,$A$188:$G$196,7,FALSE)'

# CLOSE
sheet['cls_nr'][1:-2, 5].formula = '=VLOOKUP(D161,$A$188:$G$196,7,FALSE)'

"""
DATA TABLE FOR CHARTS & VISUALS
-Total hours by role
-Total hours by stage

*other not yet worked out*
-FTE by stage
-FTE by role
-FTE overlay
-High level project timeline
"""

"""
CHARTS DATA
"""
# NAMED RANGE - Chart data
sheet['A200:G209'].name = 'cht_nr'

# INSERT HEADER
sheet['cht_nr'][0,3:].value = ['ROLE', 'HOURS', 'COST', 'RATE']

# GET ROLES
cht_list = role_rate_loc[1:,0].value
cht_role_list = cht_list[0:8]

cht_role_list

# GET RATES
cht_rate = role_rate_loc[1:,6].value
cht_rate_list = cht_rate[0:8]

cht_rate_list

# INSERT & TRANSPOSE *ROLE* DATA -> to new table (later used for charts and data viz)
sheet['cht_nr'][1:,3].options(transpose=True).value = cht_role_list

# INSERT & TRANSPOSE *RATE* DATE
sheet['cht_nr'][1:,6].number_format = '$0.00' # sets formatting before insert
sheet['cht_nr'][1:,6].options(transpose=True).value = cht_rate_list


# SUMIF() function in excel to SUM hours by role -> populate table values
sheet['cht_nr'][1:9,4].formula = '=SUMIF(D1:D179,D201,E1:E179)'

# Set cost # format also
sheet['cht_nr'][1:,5].number_format = '$0.00'

# FORMULA to calculate cost
sheet['cht_nr'][1:9,5].formula = ('=E201*G201')


# DEV set PS row to red
sheet['cht_nr'][8,3:].font.color = '#E21A23'


# BORDERS & FORMATTING
sheet['cht_nr'][0,3:].font.bold = True 

# BORDER
sheet['cht_nr'][0:8,3:].api.get_border(which_border=k.border_top).weight.set(4)
sheet['cht_nr'][0:8,3:].api.get_border(which_border=k.border_left).weight.set(4)
sheet['cht_nr'][0:8,3:].api.get_border(which_border=k.border_bottom).weight.set(4)
sheet['cht_nr'][0:8,3:].api.get_border(which_border=k.border_right).weight.set(4)

Start here. All working above. 
Need to figure out FTEs and the sample project timeline.

# Save workbook BUT leave open
book_number = 'book01.xlsx'
wb.save(book_number)

| END OF PROGRAM |
| ----------- |
| Notes for future DEV, and reference code below |

# # DATA - Hours breakdown by role
# hrs_chart_data_loc = sheet['sow_nr'][200:210,:]
# hrs_chart_data_list = ['ROLE', 'HOURS', 'COST', 'RATE']

# hrs_chart_data_loc[:,3:5].value = hrs_chart_data_list

# # set roles in new table
# role_rate_loc[1:8,0].options(transpose=True).value = role_rate_loc[1:8,0]


# DATA - Hours breakdown by stage

# DATA - Time breakdown by stage

"""
Load overlay data and store as a dataframe
Slice new dataframe (3) dataframes
Write 3 dataframes to respective excel document location
Set formulas to dynamically calculate totals
Set formulas to dynamically calculate hours based on percentages
"""
# overlay_df = pd.read_excel('sow_input_v5.0_beta.xlsx', sheet_name='OVERLAY') 

# overlay_df

# """    
# OVERLAYS
# -Scribe (BA)
# -BPC
# -EM
# -TA (formerly SA)
# """

# # Set color variable
# overlay_color = '#F9F7ED'

# """ 
# Define
# -BA and EM
# """
# scribe_ovy = overlay_df.iloc[0:1,1:]
# sheet['def_nr'][16,1:].options(index=False, header=False).value = scribe_ovy
# sheet['def_nr'][16,4].formula = '=ROUNDUP(SUM(E41:E55)*.05,0)' # Important - use ROUNDUP()

# em_ovy = overlay_df.iloc[2:3,1:]
# sheet['def_nr'][17,1:].options(index=False, header=False).value = em_ovy
# sheet['def_nr'][17,4].formula = '=ROUNDUP(SUM(E41:E55)*.05,0)' # Important - use ROUNDUP()

# sheet['def_nr'][16:18,1:].color = overlay_color

# """ 
# Design
# -BPC and EM
# """
# bpc_ovy = overlay_df.iloc[1:2,1:]
# sheet['des_nr'][16,1:].options(index=False, header=False).value = bpc_ovy
# sheet['des_nr'][16,4].formula = '=ROUNDUP(SUM(E41:E55)*.05,0)' # Product sum from define phase is the scalar

# sheet['des_nr'][17,1:].options(index=False, header=False).value = em_ovy
# sheet['des_nr'][17,4].formula = '=ROUNDUP(SUM(E61:E75)*.05,0)' # Product sum from define phase is the scalar

# sheet['des_nr'][16:18,1:].color = overlay_color

# """
# Execute
# -BPC, TA, and EM
# """ 
# sheet['bld_nr'][15,1:].options(index=False, header=False).value = bpc_ovy
# sheet['bld_nr'][15,4].formula = '=ROUNDUP(SUM(E81:E94)*.05,0)'

# sa_ovy = overlay_df.iloc[3:4,1:]
# sheet['bld_nr'][16,1:].options(index=False, header=False).value = sa_ovy
# sheet['bld_nr'][16,4].formula = '=ROUNDUP(SUM(E81:E94)*.05,0)'

# sheet['bld_nr'][17,1:].options(index=False, header=False).value = em_ovy
# sheet['bld_nr'][17,4].formula = '=ROUNDUP(SUM(E81:E94)*.05,0)'

# sheet['bld_nr'][15:18,1:].color = overlay_color

# """
# Transition
# -TA and EM
# """
# sheet['trn_nr'][16,1:].options(index=False, header=False).value = sa_ovy
# sheet['trn_nr'][16,4].formula = '=ROUNDUP(SUM(E101:E115)*.05,0)'

# sheet['trn_nr'][17,1:].options(index=False, header=False).value = em_ovy
# sheet['trn_nr'][17,4].formula = '=ROUNDUP(SUM(E101:E115)*.05,0)'

# sheet['trn_nr'][16:18,1:].color = overlay_color

# """
# Close
# #### Nothing to add at the moment ####

# """
# # Close
# # -Think about using f strings to dynamically capture percentage adjustements
# # """
# # cls_ovy = overlay_df.iloc[2:,1:]
# # sheet['cls_nr'][15:18,1:].options(index=False, header=False).value = cls_ovy
# # sheet['cls_nr'][15,4].formula = '=ROUNDUP(SUM(E58+E78+E98+E118)*.05,0)' # PM (define, design, execute, transition)
# # sheet['cls_nr'][16,4].formula = '=ROUNDUP(SUM(E98+E118)*.05,0)' # SA (execute, transition)
# # sheet['cls_nr'][17,4].formula = '=ROUNDUP(SUM(E98)*.05,0)' # BPC (execute)
# # sheet['cls_nr'][17,1:].color = overlay_color


"""
Verify active excel workbook and sheet
"""
# print(xw.apps.active)
# print(xw.books.active)
# print(xw.sheets.active)

# # Set named range for totals
# sheet['D140:G144'].name = 'tot_nr'

# # Assign values to cells by passing lists to the excel doc
# sheet['tot_nr'][1,:].value = ['TOTAL HOURS', 0, 'TOTAL COST', '$0.00']
# sheet['tot_nr'][2,:].value = ['', 'OTD', '0%', '$0.00']
# sheet['tot_nr'][3,:].value = ['', '', 'GRAND TOTAL', '$0.00']

# # Set font bold and color
# sheet['tot_nr'][1:4,:].font.bold = True
# sheet['tot_nr'][1:4,:].font.color = '#000000'

# # Set formulas
# sheet['tot_nr'][1,1].formula = '=(E139)'
# sheet['tot_nr'][1,3].formula = '=(G139)'
# sheet['tot_nr'][2,3].formula = '=(F142*G141)'
# sheet['tot_nr'][3,3].formula = '=(G141-G142)'



# # td_range = sheet['tot_nr'][1:,1:4].options(expand='table')
# td_range = sheet['tot_nr'].options(expand='table')
# td_value = td_range.value
# td_value

# """
# IMPORTANT - MacOS
# Use appscript k function as a workaround because I use a macOS operating system
# """ 
# from appscript import k

# # Set totals border(s)
# sheet['tot_nr'][1:4,:].api.get_border(which_border=k.border_top).weight.set(4)
# sheet['tot_nr'][1:4,:].api.get_border(which_border=k.border_left).weight.set(4)
# sheet['tot_nr'][1:4,:].api.get_border(which_border=k.border_bottom).weight.set(4)
# sheet['tot_nr'][1:4,:].api.get_border(which_border=k.border_right).weight.set(4)

# # Set line style
# sheet['tot_nr'][1:4,:].api.get_border(which_border=k.border_top).line_style.set(1)


# """
# Set borders on overalys 
# """
# # Define
# sheet['def_nr'][16:18,1:].api.get_border(which_border=k.border_top).weight.set(2)
# sheet['def_nr'][16:18,1:].api.get_border(which_border=k.border_left).weight.set(2)
# sheet['def_nr'][16:18,1:].api.get_border(which_border=k.border_bottom).weight.set(2)
# sheet['def_nr'][16:18,1:].api.get_border(which_border=k.border_right).weight.set(2)

# sheet['def_nr'][16:18,1:].api.get_border(which_border=k.border_top).line_style.set(1)

# # Design
# sheet['des_nr'][16:18,1:].api.get_border(which_border=k.border_top).weight.set(2)
# sheet['des_nr'][16:18,1:].api.get_border(which_border=k.border_left).weight.set(2)
# sheet['des_nr'][16:18,1:].api.get_border(which_border=k.border_bottom).weight.set(2)
# sheet['des_nr'][16:18,1:].api.get_border(which_border=k.border_right).weight.set(2)

# sheet['des_nr'][16:18,1:].api.get_border(which_border=k.border_top).line_style.set(1)


# # Build
# sheet['bld_nr'][15:18,1:].api.get_border(which_border=k.border_top).weight.set(2)
# sheet['bld_nr'][15:18,1:].api.get_border(which_border=k.border_left).weight.set(2)
# sheet['bld_nr'][15:18,1:].api.get_border(which_border=k.border_bottom).weight.set(2)
# sheet['bld_nr'][15:18,1:].api.get_border(which_border=k.border_right).weight.set(2)

# sheet['bld_nr'][15:18,1:].api.get_border(which_border=k.border_top).line_style.set(1)


# # Transition
# sheet['trn_nr'][16:18,1:].api.get_border(which_border=k.border_top).weight.set(2)
# sheet['trn_nr'][16:18,1:].api.get_border(which_border=k.border_left).weight.set(2)
# sheet['trn_nr'][16:18,1:].api.get_border(which_border=k.border_bottom).weight.set(2)
# sheet['trn_nr'][16:18,1:].api.get_border(which_border=k.border_right).weight.set(2)

# sheet['trn_nr'][16:18,1:].api.get_border(which_border=k.border_top).line_style.set(1)



# """
# Build a math checker
# -Do something later like IF 'GRAND TOTAL' == 'GRAND TOTAL' THEN font.color=green ELSE font.color=red
# """
# # Set named range for totals
# sheet['D145:G149'].name = 'chk_nr'

# # Assign values to cells by passing lists to the excel doc
# sheet['chk_nr'][0,0].value = 'MATH CHECKER'
# sheet['chk_nr'][1,:].value = ['TOTAL HOURS', 0, 'TOTAL COST', '$0.00']

# # Set color on math checker field
# sheet['chk_nr'][0,0].font.color = '#E21A23'

# # Set font bold and color
# sheet['chk_nr'][0:2,:].font.bold = True
# sheet['chk_nr'][1:2,:].font.color = '#000000'

# # Set formulas
# sheet['chk_nr'][1,1].formula = '=SUM(E18+E38+E58+E78+E98+E118+E138)' # Total hours
# sheet['chk_nr'][1,3].formula = '=SUM(G18+G38+G58+G78+G98+G118+G138)' # Total cost

# # Set border and line style
# sheet['chk_nr'][1:2,:].api.get_border(which_border=k.border_top).weight.set(4)
# sheet['chk_nr'][1:2,:].api.get_border(which_border=k.border_left).weight.set(4)
# sheet['chk_nr'][1:2,:].api.get_border(which_border=k.border_bottom).weight.set(4)
# sheet['chk_nr'][1:2,:].api.get_border(which_border=k.border_right).weight.set(4)

# # Default line style is (1) and some weird error happened. Leaving out.

# # Import INPUT dataframes
# env_df = pd.read_excel('sow_input_v5.0_beta.xlsx', sheet_name = 'ENVISION')
# int_df = pd.read_excel('sow_input_v5.0_beta.xlsx', sheet_name = 'INITIATE')
# # Skip DEFINE for now
# des_df = pd.read_excel('sow_input_v5.0_beta.xlsx', sheet_name = 'DESIGN')
# bld_df = pd.read_excel('sow_input_v5.0_beta.xlsx', sheet_name = 'BUILD')
# trn_df = pd.read_excel('sow_input_v5.0_beta.xlsx', sheet_name = 'TRANSITION')
# cls_df = pd.read_excel('sow_input_v5.0_beta.xlsx', sheet_name = 'CLOSE')

# sheet['env_nr'][1:,1:].options(index=False, header=False).value = env_df
# sheet['int_nr'][1:,1:].options(index=False, header=False).value = int_df
# # Skip DEFINE
# sheet['des_nr'][1:,1:].options(index=False, header=False).value = des_df
# sheet['bld_nr'][1:,1:].options(index=False, header=False).value = bld_df
# sheet['trn_nr'][1:,1:].options(index=False, header=False).value = trn_df
# sheet['cls_nr'][1:,1:].options(index=False, header=False).value = cls_df

# Highlight "BREAK INSERT HERE" - This is where SOW input data is injected
# sheet['bld_nr'][4,1:].color = '#ffff00'
# sheet['def_nr'][4,1:].color = '#ffff00'

# """
# RATES & ROLES
# """
# # Set roles rates named range
# sheet['A150:G159'].name = 'rr_nr'

# # Read in ROLES_RATES df 
# rr_df = pd.read_excel('sow_input_v5.0_beta.xlsx', sheet_name = 'ROLES_RATES')

# rr_df = rr_df.iloc[0:7,:]

# rr_df

# # Assign header value in excel
# rr_header_list = ['ROLE', 
#                   'ROLE NAME', 
#                   'ROLE DESCRIPTION', 
#                   'STD RATE', 
#                   'DISCOUNT', 
#                   'DISCOUNT %', 
#                   'RATE']

# sheet['rr_nr'][0,:].value = rr_header_list

# # Insert rows rates dataframe 
# sheet['rr_nr'][1:,0:].options(index=False, header=False).value = rr_df

# # Set 0% and $0.00 formatting
# sheet['rr_nr'][1:,3].number_format = '$0.00' # STD RATE
# sheet['rr_nr'][1:,4].number_format = '$0.00' # DISCOUNT
# sheet['rr_nr'][1:,5].number_format = '0.0%' # DISCOUNT %
# sheet['rr_nr'][1:,6].number_format = '$0.00' # NEGOTIATED RATE

# # Set formula to calculate final rate from discount (discount applied manually by user)
# sheet['rr_nr'][1:8,6].formula = '=(D151-E151)'

# """
# Set discount formula
# -Important for usability by non-technical user
# """ 

# # Set formula to calculate discount perecentage
# sheet['rr_nr'][1:8,5].formula = '=(E151/D151)'

# """ 
# Format roles and rates table in excel doc
# """
# # Set bold header
# sheet['rr_nr'][0,:].font.bold = True

# # Set border and line style
# sheet['rr_nr'][0:8,:].api.get_border(which_border=k.border_top).weight.set(4)
# sheet['rr_nr'][0:8,:].api.get_border(which_border=k.border_left).weight.set(4)
# sheet['rr_nr'][0:8,:].api.get_border(which_border=k.border_bottom).weight.set(4)
# sheet['rr_nr'][0:8,:].api.get_border(which_border=k.border_right).weight.set(4)

# # AGAIN default line style is (1) and some weird error happened. Leaving out.

# # Add define baseline tasks
# def_df = pd.read_excel('sow_input_v5.0_beta.xlsx', sheet_name = 'DEFINE')
# sheet['def_nr'][1:,1:].options(index=False, header=False).value = def_df

# # Use VLOOKUP() to auto-apply rates or adjusted rates from excel doc roles and rates table

# """
# Envision
# """ 
# sheet['env_nr'][1:3, 5].formula = '=VLOOKUP(D3,$A$150:$G$157,7,FALSE)'

# """
# Initiate
# """ 
# sheet['int_nr'][1:9, 5].formula = '=VLOOKUP(D21,$A$150:$G$157,7,FALSE)'

# """
# Define
# -Leave out for now
# """ 
# sheet['def_nr'][1:5, 5].formula = '=VLOOKUP(D41,$A$150:$G$157,7,FALSE)'

# """
# Design
# """ 
# sheet['des_nr'][1:4, 5].formula = '=VLOOKUP(D61,$A$150:$G$157,7,FALSE)'

# """
# Build
# """ 
# sheet['bld_nr'][1:5, 5].formula = '=VLOOKUP(D81,$A$150:$G$157,7,FALSE)'

# """
# Transition
# """ 
# sheet['trn_nr'][1:13, 5].formula = '=VLOOKUP(D101,$A$150:$G$157,7,FALSE)'

# """
# Close
# """ 
# sheet['cls_nr'][1:6, 5].formula = '=VLOOKUP(D121,$A$150:$G$157,7,FALSE)'

# # Overlays
# sheet['def_nr'][16:18, 5].formula = '=VLOOKUP(D56,$A$150:$G$157,7,FALSE)'
# sheet['des_nr'][16:18, 5].formula = '=VLOOKUP(D76,$A$150:$G$157,7,FALSE)'
# sheet['bld_nr'][15:18, 5].formula = '=VLOOKUP(D95,$A$150:$G$157,7,FALSE)'
# sheet['trn_nr'][16:18, 5].formula = '=VLOOKUP(D116,$A$150:$G$157,7,FALSE)'

# # Save workbook then close workbook (manually)
# book_number = 'book01.xlsx'
# wb.save(book_number)
# # wb.close() # Nice feature to save and close

#### All working above this lien
-Baseline template completed

# # Read in ITSM define
# itsm_df = pd.read_excel('sow_input_v5.0_beta.xlsx', sheet_name = 'ITSM')

# itsm_df.head()


# """
# Slicing (filtering itsm dataframe based on license entitlement)

# """
# # Slice dataframe using boolean indexing (like auto-filter in excel)
# itsm_lic_ent = 'standard' # Set license entitlement one (1) time here
# itsm_filt = (itsm_df[itsm_lic_ent] == 'Y')
# itsm_filt_df = itsm_df.loc[itsm_filt, :]

# itsm_filt_df


# itsm_def_df = itsm_filt_df.iloc[:,1:5] # Important start from task column

# itsm_def_df

# # Read in ITSM build
# itsm_exe_df = pd.read_excel('sow_input_v5.0_beta.xlsx', sheet_name = 'ITSM_BUILD')
# itsm_exe_df

# # Slice dataframe
# # itsm_lic_ent = 'standard'
# itsm_filt_bld = (itsm_exe_df[itsm_lic_ent] == 'Y')
# itsm_filt_bld_df = itsm_exe_df.loc[itsm_filt_bld, :]

# itsm_bld_df = itsm_filt_bld_df.iloc[:,1:5] # Important start from feature columns

# itsm_bld_df

"""
Insert new rows to create space for workshop and build tasks / features
"""
# sheet['def_nr'][5:16,:].insert(shift='down', copy_origin='format_from_left_or_above')
# sheet['bld_nr'][6:16,:].insert(shift='down', copy_origin='format_from_right_or_below')

# Set borders top, left, right, bottom
# sheet.range('B150:G160').value = 'sudz'
# sheet.range('B150:G160').api.get_border(which_border=k.border_top).weight.set(3)
# sheet.range('B150:G160').api.get_border(which_border=k.border_left).weight.set(3)
# sheet.range('B150:G160').api.get_border(which_border=k.border_right).weight.set(3)
# sheet.range('B150:G160').api.get_border(which_border=k.border_bottom).weight.set(3)

"""
https://learn.microsoft.com/en-US/dotnet/api/microsoft.office.interop.excel.xllinestyle?view=excel-pia
"""
# Line style examples
# sheet.range('B150:G160').api.get_border(which_border=k.border_top).line_style.set(1) # Continuous line
# sheet.range('B150:G160').api.get_border(which_border=k.border_top).line_style.set(-4115) # Dashed line
# sheet.range('B150:G160').api.get_border(which_border=k.border_top).line_style.set(4) # Alternating dashes and dots
# sheet.range('B150:G160').api.get_border(which_border=k.border_top).line_style.set(5) # Dash followed by two dots
# sheet.range('B150:G160').api.get_border(which_border=k.border_top).line_style.set(-4118) # Dotted line
# sheet.range('B150:G160').api.get_border(which_border=k.border_top).line_style.set(-4119) # Double line
# sheet.range('B150:G160').api.get_border(which_border=k.border_top).line_style.set(-4142) # No line
# sheet.range('B150:G160').api.get_border(which_border=k.border_top).line_style.set(13) # Slanted dashes



# # Save workbook then close workbook (manually)
# book_number = 'book18.xlsx'
# wb.save(book_number)
# wb.close() # Nice feature to save an close

# # Reopen workbook
# wb = xw.Book(book_number)
# sheet = wb.sheets[0] 

"""
Different ways to create a dataframe refresher

Using a list
List of lists
Dictionary
List of dictionaries
Zip function
Dictionary of series
""" 

# List
# lst = [1,2,3,4,5]
# df = pd.DataFrame(lst)
# df

# List of lists
# lst = [[1,2,3,4,5], [10,20,30,40,50]]
# df = pd.DataFrame(lst)
# df

# List of lists 2
# lst = [['Anna', 10], ['Bob', 20], ['Charlie', 30]]
# df = pd.DataFrame(lst)
# df

# Dictionary
# data = {'Name':['Anna', 'Bob', 'Charlie'],
#         'Score': [95, 74, 56]
#         }

# df = pd.DataFrame(data)
# df

# # List of dictionaries
# lst = [{'a':'blue', 'b':2, 'c':3},
#        {'a':45, 'b':89, 'c':90},
#        {'a':45, 'b':89}
#        ]

# df = pd.DataFrame(lst)
# df

# Zip function
# Dictionary of series

# def scribe(nr, ):
#     sheet[nr].last_cell.row, sheet[nr].last_cell.column

# nr_last_row = sheet['env_nr'].last_cell.row
# nr_last_column = sheet['env_nr'].last_cell.columns

# sheet['env_nr'][1:,0:]


# scribe_row = sheet['env_nr'].last_cell.row-2
# scribe_row


# sheet['env_nr'][scribe_row,0].color = '#704214'

#### More base formatting can go here
Think things like formatting columns and formulas
FORMULAS overlays

# Last cell practice
# print(sheet['env_nr'].last_cell.column)
# print(sheet['int_nr'].last_cell.column)
# print(sheet['def_nr'].last_cell.column)
# print(sheet['env_nr'].last_cell.row)
# print(sheet['int_nr'].last_cell.row)
# print(sheet['def_nr'].last_cell.row)

# sheet['def_nr'].last_cell.row, sheet['def_nr'].last_cell.columns

#### Read in data

# env_df = pd.read_excel('sow_input_v4.0_beta.xlsx', sheet_name='ENVISION')  
# int_df = pd.read_excel('sow_input_v4.0_beta.xlsx', sheet_name='INITIATE')    
# des_df = pd.read_excel('sow_input_v4.0_beta.xlsx', sheet_name='DESIGN')
# trn_df = pd.read_excel('sow_input_v4.0_beta.xlsx', sheet_name='TRANSITION')
# cls_df = pd.read_excel('sow_input_v4.0_beta.xlsx', sheet_name='CLOSE')

# sheet['env_nr'][1,1].options(index=False, header=False).value = env_df
# sheet['int_nr'][1,1].options(index=False, header=False).value = int_df
# sheet['des_nr'][1,1].options(index=False, header=False).value = des_df
# sheet['trn_nr'][1,1].options(index=False, header=False).value = trn_df
# sheet['cls_nr'][1,1].options(index=False, header=False).value = cls_df

# sheet.range('G4').number_format = '$0.00'

# sheet.range('A3').api.entirerow.insert()

# empty_row_data = ['','','','','','','']

# sheet['env_nr'][1,5].value = '$0.00'
# sheet['int_nr'][1,5].value = '$0.00'
# sheet['def_nr'][1,5].value = '$0.00'
# sheet['des_nr'][1,5].value = '$0.00'
# sheet['bld_nr'][1,5].value = '$0.00'
# sheet['trn_nr'][1,5].value = '$0.00'
# sheet['cls_nr'][1,5].value = '$0.00'
