==> app_SOWmAcHiNe_2019/2_overlays.ipynb <==
import xlwings as xw
from openpyxl import Workbook, load_workbook
import xlsxwriter 

import pandas as pd
import numpy as np
import datetime as dt 
import time
import sys

"""
Latest xlwings release: v0.27.15
https://www.xlwings.org/

"""
xw.__version__ 

from appscript import k

| OVERLAYS | Add overlay roles to phases (where applicable) |
| ----------- | ----------- |

# Connect to open workbook
book_number = 'book01.xlsx'
wb = xw.Book(book_number)
sheet = wb.sheets[0]

ovy_df = pd.read_excel('sow_input_data.xlsx', sheet_name='OVERLAY') 

ovy_df

# Set color variable
ovy_color = '#F9F7ED'

"""DEFINE - 1/4 stages with overlays"""
# Set scribe
scribe_ovy = ovy_df.iloc[0:1,1:]
sheet['def_nr'][-4,1:].options(index=False, header=False).value = scribe_ovy
sheet['def_nr'][-4,4].formula = '=ROUNDUP(SUM(E41:E75)*.04,0)' # Important - use ROUNDUP()
# Set EM
em_ovy = ovy_df.iloc[2:3,1:]
sheet['def_nr'][-3,1:].options(index=False, header=False).value = em_ovy
sheet['def_nr'][-3,4].formula = '=ROUNDUP(SUM(E41:E75)*.04,0)' # Important - use ROUNDUP()

# Set color
sheet['def_nr'][-4:-2,1:].color = ovy_color

# BORDERS & FORMATTING
sheet['def_nr'][-4:-2,1:].font.bold = True 

# BORDERS
sheet['def_nr'][-4:-2,1:].api.get_border(which_border=k.border_top).weight.set(2)
sheet['def_nr'][-4:-2,1:].api.get_border(which_border=k.border_left).weight.set(2)
sheet['def_nr'][-4:-2,1:].api.get_border(which_border=k.border_bottom).weight.set(2)
sheet['def_nr'][-4:-2,1:].api.get_border(which_border=k.border_right).weight.set(2)


"""DESIGN - 2/4 stages with overlays"""
# Set BPC
bpc_ovy = ovy_df.iloc[1:2,1:]
sheet['des_nr'][-4,1:].options(index=False, header=False).value = bpc_ovy
sheet['des_nr'][-4,4].formula = '=ROUNDUP(SUM(E81:E95)*.04,0)' # Product sum from define phase is the scalar

# Set EM
# em_ovy = ovy_df.iloc[2:3,1:] # ALREADY SET IN PREVIOUS CALL
sheet['des_nr'][-3,1:].options(index=False, header=False).value = em_ovy
sheet['des_nr'][-3,4].formula = '=ROUNDUP(SUM(E81:E95)*.04,0)' # Product sum from define phase is the scalar

# Set color
sheet['des_nr'][-4:-2,1:].color = ovy_color

# BORDERS & FORMATTING
sheet['des_nr'][-4:-2,1:].font.bold = True 

# BORDERS
sheet['des_nr'][-4:-2,1:].api.get_border(which_border=k.border_top).weight.set(2)
sheet['des_nr'][-4:-2,1:].api.get_border(which_border=k.border_left).weight.set(2)
sheet['des_nr'][-4:-2,1:].api.get_border(which_border=k.border_bottom).weight.set(2)
sheet['des_nr'][-4:-2,1:].api.get_border(which_border=k.border_right).weight.set(2)

"""EXECUTE - 3/4 stages with overlays"""
# Set BPC
# bpc_ovy = ovy_df.iloc[1:2,1:] # ALREADY SET
sheet['bld_nr'][-5,1:].options(index=False, header=False).value = bpc_ovy
sheet['bld_nr'][-5,4].formula = '=ROUNDUP(SUM(E101:E134)*.04,0)' # Product sum from define phase is the scalar

# Set SA (TA)
sa_ovy = ovy_df.iloc[3:4,1:]
sheet['bld_nr'][-4,1:].options(index=False, header=False).value = sa_ovy
sheet['bld_nr'][-4,4].formula = '=ROUNDUP(SUM(E101:E134)*.04,0)'

# Set EM
# em_ovy = ovy_df.iloc[2:3,1:] # ALREADY SET
sheet['bld_nr'][-3,1:].options(index=False, header=False).value = em_ovy
sheet['bld_nr'][-3,4].formula = '=ROUNDUP(SUM(E101:E134)*.04,0)' # Product sum from define phase is the scalar

# Set color
sheet['bld_nr'][-5:-2,1:].color = ovy_color

# BORDERS & FORMATTING
sheet['bld_nr'][-5:-2,1:].font.bold = True 

# BORDERS
sheet['bld_nr'][-5:-2,1:].api.get_border(which_border=k.border_top).weight.set(2)
sheet['bld_nr'][-5:-2,1:].api.get_border(which_border=k.border_left).weight.set(2)
sheet['bld_nr'][-5:-2,1:].api.get_border(which_border=k.border_bottom).weight.set(2)
sheet['bld_nr'][-5:-2,1:].api.get_border(which_border=k.border_right).weight.set(2)

"""TRANSITION - 4/4 stages with overlays"""
# Set SA (TA)
# sa_ovy = ovy_df.iloc[3:4,1:] # ALREADY SET
sheet['trn_nr'][-4,1:].options(index=False, header=False).value = sa_ovy
sheet['trn_nr'][-4,4].formula = '=ROUNDUP(SUM(E141:E155)*.04,0)'

# Set EM
# em_ovy = ovy_df.iloc[2:3,1:] # ALREADY SET
sheet['trn_nr'][-3,1:].options(index=False, header=False).value = em_ovy
sheet['trn_nr'][-3,4].formula = '=ROUNDUP(SUM(E141:E155)*.04,0)' # Product sum from define phase is the scalar

# Set color
sheet['trn_nr'][-4:-2,1:].color = ovy_color

# BORDERS & FORMATTING
sheet['trn_nr'][-4:-2,1:].font.bold = True 

# BORDERS
sheet['trn_nr'][-4:-2,1:].api.get_border(which_border=k.border_top).weight.set(2)
sheet['trn_nr'][-4:-2,1:].api.get_border(which_border=k.border_left).weight.set(2)
sheet['trn_nr'][-4:-2,1:].api.get_border(which_border=k.border_bottom).weight.set(2)
sheet['trn_nr'][-4:-2,1:].api.get_border(which_border=k.border_right).weight.set(2)

# SAVE
book_number = 'book01.xlsx'
wb.save(book_number)

| END OF PROGRAM |
| ----------- |
