==> app_SOWmAcHiNe_2019/3_baseline_tasks.ipynb <==
| BASELINE TASKS | SOW Template Setup |
| ----------- | ----------- |

import xlwings as xw
from openpyxl import Workbook, load_workbook
import xlsxwriter 

import pandas as pd
import numpy as np
import datetime as dt 
import time
import sys

"""
Latest xlwings release: v0.27.15
https://www.xlwings.org/

"""
xw.__version__ 

from appscript import k

# Connect to workbook
book_number = 'book01.xlsx'
wb = xw.Book(book_number)
sheet = wb.sheets[0]

"""
Verify active excel workbook and sheet
"""
print(xw.apps.active)
print(xw.books.active)
print(xw.sheets.active)

# Import baseline task data from excel
base_env_df = pd.read_excel('sow_input_data.xlsx', sheet_name = 'ENVISION')
base_int_df = pd.read_excel('sow_input_data.xlsx', sheet_name = 'INITIATE')
base_def_df = pd.read_excel('sow_input_data.xlsx', sheet_name = 'DEFINE')
base_des_df = pd.read_excel('sow_input_data.xlsx', sheet_name = 'DESIGN')
base_bld_df = pd.read_excel('sow_input_data.xlsx', sheet_name = 'BUILD')
base_trn_df = pd.read_excel('sow_input_data.xlsx', sheet_name = 'TRANSITION')
base_cls_df = pd.read_excel('sow_input_data.xlsx', sheet_name = 'CLOSE')

base_env_df

# Write dataframes to SOW excel doc
sheet['env_nr'][1:,1:].options(index=False, header=False).value = base_env_df
sheet['int_nr'][1:,1:].options(index=False, header=False).value = base_int_df
sheet['def_nr'][1:,1:].options(index=False, header=False).value = base_def_df
sheet['des_nr'][1:,1:].options(index=False, header=False).value = base_des_df
sheet['bld_nr'][1:,1:].options(index=False, header=False).value = base_bld_df
sheet['trn_nr'][1:,1:].options(index=False, header=False).value = base_trn_df
sheet['cls_nr'][1:,1:].options(index=False, header=False).value = base_cls_df

# FORMULAS - QA / TESTING

# Planning
sheet['trn_nr'][3,4].formula = '=ROUNDUP(E138*.05,0)' # IMPORTANT only QA hours from BUILD phase
sheet['trn_nr'][4,4].formula = '=ROUNDUP(E138*.05,0)'

# Executing
sheet['trn_nr'][5,4].formula = '=ROUNDUP(E138*.05,0)'
sheet['trn_nr'][6,4].formula = '=ROUNDUP(E138*.05,0)'

# SAVE
book_number = 'book01.xlsx'
wb.save(book_number)

| END OF PROGRAM |
| ----------- |
