==> app_SOWmAcHiNe_2019/5_clean_up.ipynb <==
| CLEAN UP | Remove placeholder values |
| ----------- | ----------- |

import xlwings as xw
from openpyxl import Workbook, load_workbook
import xlsxwriter 

import pandas as pd
import numpy as np
import datetime as dt 
import time
import sys

"""
Latest xlwings release: v0.27.15
https://www.xlwings.org/

"""
xw.__version__ 

from appscript import k

# Connect to workbook
book_number = 'book01.xlsx'
wb = xw.Book(book_number)
sheet = wb.sheets[0]

"""
Delete Placeholders
"""

# Harcode delete CLOSE placeholders
sheet['A166:G177'].delete(shift='up')

# Harcode delete TRANSITION placeholders
sheet['A151:G155'].delete(shift='up')

# Harcode delete BUILD placeholders
sheet['A125:G134'].delete(shift='up')

# Harcode delete DESIGN placeholders
sheet['A84:G95'].delete(shift='up')

# Harcode delete DEFINE placeholders
sheet['A65:G75'].delete(shift='up')

# Harcode delete INITIATE placeholders
sheet['A29:G37'].delete(shift='up')

# Harcode delete ENVISION placeholders
sheet['A5:G17'].delete(shift='up')

# Harcode delete RATE CHARTS placeholders
sheet['A136:G136'].delete(shift='up')

# Harcode delete RATE MAIN placeholders
sheet['A124:G124'].delete(shift='up')

# SAVE
book_number = 'book01.xlsx'
wb.save(book_number)

| END OF PROGRAM |
| ----------- |
| Notes for future DEV, and reference code below |


# TEST / DEMO - hardcode drop placeholders
# for i in sheet['des_nr'][:,1]:
#     if i.value == 'XXX':
#         row_loc = i.row
#         # print(row_loc)
#         sheet['des_nr'][row_loc,:].delete(shift='up')

# Get a list of named ranges in sheet
# for name in wb.names:
#     #print(name.name, name.refers_to_range)
    
#     nr_list = name.name
#     print(nr_list)

# IMPORTANT - Remove named ranges for cleanup
# if len(wb.names) > 0:
#     #print(len(wb.names))
#     for n in range(0, len(wb.names)):
# 	    wb.names[0].delete()

# Verify there are no named ranges
# for name in wb.names:
#     print(name.name, name.refers_to_range)

# Create one (1) named range
# sheet['A2:G180'].name = 'clean_nr'

# for i in sheet['clean_nr'][:,1]:
#     if i.value == 'XXX':
#         row_loc = i.row
#         sheet['clean_nr'][row_loc,:].delete(shift='up')

# for i in sheet['env_nr'][0:,1]:
#     if i.value == 1:
#         sheet['env_nr'][i.row,:].delete()

# sheet['A1'].current_region.name = 'cleanup_rng'

# for i in sheet['cleanup_rng'][:-1,:]:
#     print(i)
    # if i.value == 'pre-sales':
    #     # delete entire row for all rows where at least one cell value == 'pre-sales'
    #     # sheet.range(i.row,1).delete()
    #     sheet['cleanup_rng'][i.row,:].delete()

# sheet['A1'].current_region.name = 'cleanup_rng'

# for i in sheet['cleanup_rng']:
#     if i.value == 'pre-sales':
#         ps_cell_row_num = i.row
#         ps_cell_col_num = i.column

# print(ps_cell_row_num)
# print(ps_cell_col_num)

# sheet['A1'].current_region.name = 'cleanup_rng'
# filter_word = 'pre-sales'

# sheet['cleanup_rng'][0:,1:3].options(ndim=2).value == ['pre-sales', 'pre-sales']


    

# cup_rng = sheet['A1'].current_region

# # ph_list = sheet['env_nr'][3,:].value

# for i in cup_rng:
#     if i.value == 'pre-sales':
#         print(i)

# # Create placeholder list from one row
# ph_list = sheet['env_nr'][3,:].value

# ph_list

# clean_filt = sheet['env_nr'][:,1:3] == 'pre-sales'

# # clean_filt

# Tooltips
# dir(sheet['A1'])

# sheet['A1'].api

# cup_rng = sheet['A1'].current_region

# for i in cup_rng:
#     if i.value == ('pre-sales' and 'placeholder'):
#         print(i)


# for i in sheet['env_nr'][-1,:]:
#     i = i + 1
#     if sheet['env_nr']('A' + str(i)).value == 'pre-sales' or 'placeholder':
#         xw.Range('A' + str(i)).clear()
#         continue
#     else:
#         continue

# for i in sheet['env_nr'][:,1]:
#     if i.value != 'pre-sales':
#         pass
#     elif i.value == 'pre-sales':
#         print(i)
        
        

# del_rng_list = []

# for i in sheet['env_nr'][:,1]:
#     if i.value == 'pre-sales':
#         del_rng_list.append(i)
               
# print(del_rng_list)


# ph_list = []

# for i in sheet['env_nr'][:,1]:
#     if i.value == 'pre-sales':
#         ph_list.append(i)
